wdt implementation

+++ src/display/utils.py
@@ -15,12 +15,13 @@ def couple_roi_names(clinical_names, target):
     return matches
 
 def clear_directory_content(PATH):
-    for filename in os.listdir(PATH):
-        file_path = os.path.join(PATH, filename)
-        if os.path.isfile(file_path) or os.path.islink(file_path):
-            os.remove(file_path)
-        elif os.path.isdir(file_path):
-            shutil.rmtree(file_path)
+    if os.path.exists(PATH):
+        for filename in os.listdir(PATH):
+            file_path = os.path.join(PATH, filename)
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.remove(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
 
 def rm_pss(s, delim='_'):
     return delim.join(c for c in s if c.isalnum())
